requirements.txt
openai>=1.0.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
yt-dlp>=2024.0.0
openai-whisper>=20231117
pyyaml>=6.0
python-dotenv>=1.0.0
ffmpeg-python>=0.2.0
httpx>=0.25.0
PyJWT>=2.8.0
pydantic>=2.0.0
click>=8.0.0
rich>=13.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0
youtube-transcript-api>=1.2.0
